jit: move CTypeContext out of the LLVM backend
Make xdsl.jit.c_type_context backend-agnostic and split the ctypes
registration helper into register_builtin_ctypes for builtin types and
register_llvm_ctypes for LLVM dialect types. to_c_func_type now takes
argument and result types directly, so the variadic check moves to the
LLVM wrapper, which is the only place an LLVMFunctionType is available.

The pyjit prototype registers both helpers on its backend instead of
mapping f64 by hand, so void and pointer types no longer fail.
